New Git push procedure - first draft
Introduce GitCompoundResult to hold GitCommandResults from several repositories.
Push by GitPusher. Display notification in case of success, error or partial success (when some repositories succeeded, but some of them failed).
GitPushSpec to hold both remote and refspec (parameters of 'git push').
GitPushInfo to hold pushed repositories (with the number of commits to push - just to show that number later in case of successful push) and the  GitPushSpec which is common for all repositories.