Hide "Push to all branches" checkbox in the push dialog.
Reason: the option is not quite clear, and may confuse users. Especially it is confusing in a multi-repo configuration (where "push all branches" might be understood as "push branches from all repositories").
It also doesn't give any visual effect if there are no changes in other branches (commit list stays the same after reloading).
The option is hidden, but not removed, because pushing all branches is default for command line 'git push'. Therefore if there is a feedback to implement the option, it will be reconsidered.