Test fixes
The problem was that FormatterUtil hold FormatterUtilHelper instances at the static field (ourHelper). JavadocFormatterUtilHelper was added to that field from static initialisation block of JavadocManagerImpl. The problem was that the later class is declared as service, i.e. is lazily instantiated. So, if target tests were executed before JavadocManagerImpl instantiation, they were failed.

1. Removed static processing for FormatterUtilHelper;
2. Merged FormatterUtilHelper with WhiteSpaceFormattingStrategy;