GitMergeOperation
Add "Merge" operation to the GitBranchPopup.
Extract more methods on "local changes overwritten by" error from GitCheckoutOperation, and reuse them in GitMergeOperation.
Add some tests on GitMergeOperation; DRY GitBranchOperationsTest a bit.
Rename GitWouldBeOverwrittenByCheckoutDialog to GitSmartOperationDialog. Make "force" button optional - it has no sense for merge.
Extract GitMergeCommittingConflictResolver from GitUpdateProcess to reuse in GitMergeOperation. Refresh the root after commit if needed (to understand that the merge has been committed; this is not needed in case of update, because the ChangeListManager is frozen and then refreshed).