Detect nested roots in folders for <Project> mapping.
GitRootConverter is called each time when ProjectLevelVcsManager#getVcsRoots is called.
There all content roots (in the case of <Project> mapping) are converted to a set of Git roots.
If <Project> mapping is defined, all Git roots are found by GitRootDetector and returned.
Otherwise, the same strategy of finding the closest Git above the given root is used.

Update tests to reflect the possibility.

Note: This is not final version, because it queries GitRootDetector each time,
which is bad for performance.