GitCommand: fetch and show are read-commands
show is obviously a reading command
fetch doesn't modify the index, so it is a read-command, since avoiding conflicts around the index.lock is the purpose of the locking policy.
push is also a read-command, but let it be "write" as an additional protection against "commit during push" problem, when a commit may be unintentionally pushed.