Enable the BuiltInServer for tests
It is needed for Git tests.
For Cucumber tests we could stay with the previous implementation,
since they are run after all other tests on our build server; but it
is not acceptable for JUnits. The alternative to start/stop
before/after each test is also unacceptable slow down.

Remove the test startup instruction from GitCucumberWorld.
Move default server port setting to a utility function.