[git] Fix equals & hashcode for GitRepoInfo
We consider GitRepoInfos as equal if, in particular,
they have the same set of remotes, while the order of these remotes
in collection is not interesting.

(Remotes are supplied as ArrayLists, while equals for the ArrayList
considers the order of the elements)

Also hashCode() is adjusted to be symmetric for collections,
to conform equals(). It became less efficient, but we don't plan to
store GitRepoInfos in HashSets.