[git] Better fix of equals & hashcode for GitRepoInfo
We consider GitRepoInfos as equal if, in particular,
they have the same set of remotes, while the order of these remotes
in collection is not interesting.
(Remotes are (or may be) supplied as ArrayLists, while equals for the
ArrayList considers the order of the elements).

Use LinkedHashSet to preserve the original ordering,
but use the needed equality definition.