[log] Support loading details in the multi-repo setup
When user navigates to an old commit (via scrolling or via clicking
to a branch reference), details for this commit & several commits
around it are loaded.

The nodes around selected node can belong to different repositories,
therefore we need to know which repository each node belongs to,
to pass the hash to `git show` invoked on the correct repository.

* Store the root in each Ref.
* Use DataGetter#findRefByNode to navigate from each node upwards to
  find the first reference on the current branch, and take the root
  value from it to identify which root does the node belong to.