[log] Protect commit hierarchy against incompatible equals()
All implementor of the CommitParents should be compared only by their
hashes, i. e. a GitCommit with a hash A is equal to the
SimpleCommitParents with the same hash A, although other fields
are not equal.

As a protection against custom equals() and hashcode implementations
that could break this rule, make CommitParents and all its
descendants classes instead of interfaces and declare equals and
hashcode final.