[log] Store the reference to the current repository root in Branch
Walking from a node to the nearest reference can be expensive
=> store the current root in Branch objects (there is not too many of
them, so it is safe in terms of memory usage).

Branch is created from another Branch, or from a leaf node,
which has a reference label on it (and VcsRef has information about
the root).
The only exception are tags, which are not provided by GitLogProvider
yet, but this is no be fixed shortly.