[log] Invalidate myLogData on refreshCompletely()
We could do "myLogData = null" directly in refreshCompletely(),
but it would lead to write-access to myLogData from different threads
(currently it is modified only from the myBackgroundQueue).

Therefore we do it from the same background queue,
which we clear on initialize. (if another refresh process is in
progress, we will still need to wait until it completes of course).