[log] Don't draw all references on the branches panel: limit by width
Otherwise repaint is slow if there are many refs.
Pass the maximum width to the RefPainter, and don't draw if further
refs would be invisible.

The problem affects only the BranchesPanel: other RefPainter clients
draw a limited amount of refs (only refs on the given commit).
They have other problems though which will be solved later.