[log] IDEA-115480 Correctly identify details of selected rows
DetailsPanel, Actions:
Don't ask the Graph about nodes or details of the selected rows,
because in the case of non-graph filter the result will be incorrect.
Instead take the value from the table model.

Let the CommitCell hold the Hash value. In some cases it might be
null (for example, when there are "empty" lines in the table
containing only graph fragments).

Let DataGetter return value for the hash, not only for the node.
This can be slow potentially, will be optimized or moved to the
background later.

Let Copy Hash action operate on Hash instead of VcsFullCommitDetails.
Make Cherry-Pick and Open on GitHub be unavailable if details of the
selected hashes were not loaded yet.