[log] invokeLater only if needed
It fixes too long collaps of a linear branch.
The performance problem happened because after the branch was collapsed
we jumped to the commit not immediately, but Later,
which happened after JTable repaint.
The repaint on the other hand tried to guess the area to repaint,
but didn't succeed because we were staying in the invalid area,
in which case repaint was called for the whole table
(see BasicTableUI.paint "if (rMin == -1)..."