[log] Allow null roots in the GraphTableModel.
Some rows may be empty (contain no commits): for instance,
in the case of graph filtered by branch, when there will be just
graph fragments drawn.

Therefore it is not an error to have a null node,
and therefore it is not an error to receive the FAKE_ROOT in the
color manager.

Supply correct root to the LoadingDetails (the root is known because
is taken from the Branch of a Node).