[log] Don't wait under a modal progress until the log is loaded & built
The action inside model progress is executed in EDT,
which may lead to a deadlock, if showFullLog makes something like
invokeAndWait (what it actually does).

Let it just do nothing and jump to the commit when the log is loaded.

A better fix would be to introduce some less-modal progress like
JBLoadingPane (and use it for other possibly long running operations
as well).