[git] IDEA-109511 Quick-fix of "Git Add" for tracked files.
The problem was introduced in 6aa542e: the common
ScheduleForAdditionAction doesn't know that there are VCSs which can
ADD not-unversioned files.

Correct fix would be to introduce some API which plugins could use to
specify if the file of given status can be added or not.
But the timing doesn't allow to do it now.

Therefore partly return to a separate GitAdd action, but this time
extend ScheduleForAdditionAction to follow the common process to add
files.
Let specify a function determining if the file status is suitable for
calling "add to vcs" on it.

This fix keeps IDEA-76670 & IDEA-104070 fixed, but reopens IDEA-98689.
(cherry picked from commit 645e5c2)