[git] IDEA-116690 Fix "Select in Git Log" from File History
It used to navigate to the old log, therefore nothing happened with
new log implementation.

* Refer to new VcsLog from SelectRevisionInGitLogAction.
  Ask VcsLogUi to get the VcsLog instance, because:
  - VcsLog is not available in the DataContext of this action;
  - VcsLogImpl can't be easily made a Service, because it depends on
    non-components (this is to be changed later).
* VcsLogManager#getLogUi can be null if the log is not initialized
  yet => fix annotation, add javadocs.
* Update the usages of these getters to be aware of pre-init condition
  in SelectRevisionInGitLogAction & VcsLogQuickSettingsActions.
* Make the "Log" tab name public & move it to more appropriate place:
  VcsLogContentProvider.

TODO:
If the log was not initialized yet when user invoked the action,
the tab will be open, and the log will be shown,
but the commit won't be selected.
This is a bug, but not fixing because this worked the same
way previously, and the fix is not trivial.
To be fixed later.
(cherry picked from commit ce447db)