[log] Separate loading indicator for "Contained in branches"
Details are loaded in a batch and are preloaded on start and in
certain cases, while "contained in branches" information need to be
queried for each commit individually.

To not make user wait to see details while "contained in branches" is
loading, show a separate small loading icon for this process.

(cherry picked from commit 7c6bb198ae241f0854dc4963b4dae269b6bc8d9e)