[git] DRY
'git log --no-walk <hashes>' behaves the same as 'git show <hashes>'
because --no-walk tells git "Only show the given commits,
but do not traverse their ancestors".

'unsorted' tells Git to return commit details in the same order as
hashes passed to the command.
Among current usages of the method, we don't care about the order,
but seems correct for possible future methods relying on this.