[vcs] Don't wrap FilePath creation into read action
There is no point in this read action, but it takes the lock => may
slow down if the utility method is used during a large VCS process
(even if taken atomically).

VCS requires file content rarely, so the only benefit from this read
was in the situation, if a VirtualFile gets created when we've
already went to createForDeletedFile (and similar situations).
However, it is not a big deal, especially since VirtualFile is
refreshed inside the FilePathImpl.