IDEA-82129 diff: fix synchronized scrolling in merge dialog for left-middle panels
Class was written in assumption, that diff for panels is counted from left to right (the one that left - has index 0 and the one that right has index 1).
But this assumption wrong for merge, because both diffs counted with base revision on the left.
Thus previously left-middle panels was aligned in wrong order.

P.S. Table of corresponding Master/Slave pairs for merge ScrollingContext
  master = SIDE?.getIndex() + (+?)
  slave = SIDE?..otherSide().getIndex() + (+?)

LEFT [0]
  SIDE1 - (+0)  ->   m - 0, s - 1
  SIDE1 - (+1)  ->   m - 1, s - 2
MIDDLE [1]
  SIDE2 - (+0)  ->   m - 1, s - 0
  SIDE1 - (+1)  ->   m - 1, s - 2
RIGHT [2]
  SIDE2 - (+1)  ->   m - 2, s - 1
  SIDE2 - (+0)  ->   m - 1, s - 0