Add several changes in "Tasks and Contexts" subsystem API to avoid previous problems with issues missing in "Open Task" dialog (e.g. see IDEA-110012).
These problems stem from the fact that previously closed issues were filtered on the client side. E.g. consider that server has 15 open and 5 closed issues, so with default page size of 10, it's expected that without pattern and excluding closed issues "Open Task" dialog will contain two pages (10 and 5). First time we attempt to fetch at most 10 issues from server (max = 10). If it happens now that response contains 5 open and 5 closed issues, only 5 open will be added to the list, the latter won't be filled up and thus "..." item, that is necessary for initiating new request with the limit of 20 issues, won't be displayed either. It's indeed very common scenario for a large tracker with huge number of closed issues

So the following changes were made:

* TaskRepository#getIssues signature was changed to support pagination and filtering closed issues on server side. TaskSearchSupport and TaskManager was updated accordingly.
 * Because some trackers still don't have aforementioned capabilities, ChooseByName dialog was changed to allow displaying "..." item at the bottom regardless of whether the list was filled up or not.