[log] IDEA-121992 Recalculate old heads values
The VcsLogBranchFilter used to store precalculated values of commit
indices of its heads.
However, is it likely that after refresh these values change
=> setFilter will get either non-head commits, or even invalid (if
head commit was rebased/amended).

Don't precalculate these values, perform calculation each time, which
is fast because there are not too many head ref in the filter
(usually 0, 1 or 2).