[log] remove dependency
VcsLogUserRegistry interns values on addUser => the only drawback on
removing of createUser() method is that many VcsUserImpl objects will
be temporary created, but all dups will be removed right after
they are added to the registry.

On the other hand we remove this not very nice dependency on the
VcsLogManager.