IDEA-119996 FilePathImpl: avoid VFS refresh from getVirtualFile etc.
Actually, a revert of e31663b & ee19cc4 & 57c4367.

There is the public refresh() methods for those who needs to refresh.
Otherwise getPath(), isDirectory() and getVirtualFile() itself are
accessed from too many places and too often, and clients are not aware
that it may be costly to call this method frequently.