BaseAnalysisAction now respects ignored directores.
When traversing the files looking for candidates BaseAnaysisAction
would blindly traverse all directories and all files below any
directory contained in the project, disregarding subdirectories that
have been marked as excluded. For projects that contain large
subdirectories mapped to network devices or joined, caching file
systems, this is prohibitively expensive; sometimes appearing to
hang IDEA on such things as right-clicking in the project view.

With this change BaseAnalysisAction will no longer traverse the
content of directories that have been excluded from the project.