[git] IDEA-124821 Don't make the auto-generated changelist default
It is not needed: commit dialog can easily commit a non-default
changelist. On the other hand, making the auto-changelist non-default
will help to avoid such problems as IDEA-103094, when the changelist
is not removed after cherry-pick for some reason and then is used as
the default one.

This also should help with IDEA-120133.