Adds the ability for secondary toolbar buttons to be right aligned.
This ability must be turned on via the 'Registry' to have any effect.
Otherwise, secondaries will continue to function as they do now.
It also must be opted into for each toolbar as secondaries in other toolbars use the cogwheel dropdown.
Right aligned secondaries are placed to the right of the search button.

This right aligning is required in Android Studio to right align our Google Login widget -- which is a
requirement for common look and feel with other applications that support login.