[vcs] Preselect nothing in ChangesTreeList by default
On Show Diff, the ChangesTreeList limits files shown in diff to the
selected folder.

Thus if the first row is preselected (as it were before), and commit
changes files in different folders, automatic Show Diff from the log
table doesn't show everything, but only files from the first root
node which is confusing.

=> if there is nothing to select by default, don't select the first row.