WEB-13675 Do not insert redundant alignment between array elements and properties.
This alignment included in ChildAttributes for new properties and array elements
leads to spaces being inserted instead of tabs inside indents for corresponding
blocks when both SMART_TABS and USE_TAB_CHARACTER options are active.

Test data for tests of line movements was updated because it made use of
these alignments.