PY-13537 Use configurable list of ignored classes instead of checkbox
We've decided that possibility to suppress warnings about bad
method names for *all* standard classes is too general and may be
non-obvious for user. After all if inspections is too bothering it's
simpler to disable it at all. Much better way is too allow to
configure black-list of concrete classes for descendants of which
such warnings will be suppressed (as already done for some other
inspections in PyCharm).