PY-14158 Remove PyUnresolvedReferenceInspection from inspection-black-list.txt
It turned out that PyUnresolvedReferenceInspection was indeed initially
included inspection-black-list.txt. But at some moment it was moved to
subpackage and that broke its settings. Because we already use new
serialization format for this inspection (see fix for PY-14359), it
doesn't make sense to keep this class in the blacklist anymore.
(cherry picked from commit 614769c)