[log] Clarify branch comparison logic & API
Branches layout (and importance) is different from the order of
labels in the commit renderer => use different comparators for this.

For example, HEAD is the most important label and should be displayed
at the first position; but it is the less important branch head
identifier, because we want to keep layout when possible while HEAD
is constantly moving (especially during branch checkout and rebase).