Fixed callable inspection for usages-based structural types
We suppose that a structural type inferred from the usages of
attributes is callable, it's just another aspect of its attribute-based
contract.

We don't add '__call__' to the type at the moment, but it might be a
better solution. Some other special methods like '__getitem__' are
processed correctly because they are resolvable references to
appropriate methods.