Don't collect attributes conditionally accessed inside 'if' statements
It's a workaround for false positives in type checking if a function
expects its parameter to be of various types depending on some
condition we couldn't analyze statically. Another possible source of
false positives is try-except statements, but they are less likely, so
we ignore them at the moment hoping that benefits from collecting
the attributes of a parameter inside try-except will beat losses in
the amount of false positives.