Fall back to resolving super class expression if its type is not PyClassLikeType (PY-11541)
This chanage fixes a regression that was introduced in the fix to
PY-11541 for AST-based 'str' class in __builtins__.py.

Currently 'basestring' has type 'str | unicode' for Python 3. This means
that it is not added to the list of the class-like super types of 'str'.
Now we try to resolve super class exrpressions to classes as a fall
back mechanism and return the types of these classes.