PY-11932 Put nulls for unresolved class types in MRO
Some of the ancestor class-like types may be unresolved. The best we
can do is to put nulls for all the unresolved classes to appropriate
places inside the MRO chain. What we used to do in this case it return
the empty sequence of the ancestors.

(cherry picked from commit b37b0440d06989a4612262a1574678ffa8981650)