[git] IDEA-81563 Correctly identify Git root for submodule changes
When the submodule is changed in the super-project, the root of this
change is this super-project, not the submodule root.

In particular, this led to exception when trying to merge conflicts
when one of conflicts was the submodule.

(cherry picked from commit 169bf44357db5d28833ea4a74767e9eb5d6849a9)