Fall back to old-style ancestors if C3 MRO fails and there are unresolved ancestors (PY-11401)
If the C3 MRO algorithm fails and the old-style ancestors algorithm
tells that there are unresolved ancestors, we return a single 'null'
ancestor as a sign that we don't know who the ancestors really are.

If there are no unresolved ancestors, then we fall back to the
old-style ancestors algorithm in order to make resolve work. A future
inspection for detecting incorrect MRO may warn the user about this
situation.