If mro() is overridden in metaclass, use normal MRO + unresolved ancestor (PY-11401)
We cannot evaluate the result of an overridden mro() method, so we
don't know the actual MRO chain. Let's assume that MRO is almost the
same as without this override and add an element of uncertainty by
appending a fake unresolved ancestor type to the MRO chain. Doing so
results in, for example, the unresolved references inspection ignoring
unresolved references for such a class.

(cherry picked from commit 645bd3a1b27bb3c3a45e557350fb3b81da1ab52a)