[kb] Fix for IdeaSettingsTest that fails with the 139 branch.
This is not actual for the master branch, because the IdeSettingsDialog class is replaced with another class and will be removed soon.
To ensure that the test is not failing after the fix, you could set the following registry key to false.
ide.new.settings.view=false

I'll backport the fix to the 139 branch after review.
(cherry picked from commit 1fb3ee9)