check myAssertThreading before using assertIsDispatchThread
The DocumentImpl constructor takes an argument forUseInNonAWTThread, which is stored in the myAssertThreading field. While this field is checked before some assertion, it is not checked before others. This change checks the field before all assertIsDispatchThread calls, enabling users to use DocumentImpl outside the dispatch thread (as indicated by the constructor parameter).