Fixed SOE in PyClass.isOverriddenMRO() check (PY-15002)
In Python 2 it is possible to create an endless metaclass chain by
defining '__metaclass__' at the top-level to be one of the classes
defined in the module. We assume that these circular metaclasses don't
override `mro()` in order to prevent infinite recursion in this check.
(cherry picked from commit 332709f)