Prevent exponential growth of union types of PyJavaClassType (PY-15055)
For Java method references in the context of Jython multi-resolve
returns all the overloaded methods. Return types of these methods are
joined as a union type. If there is no equality for PyJavaClassType,
then the chain of builder-style calls to overloaded methods results in
exponential growth of the union type.