Prohibit `rename` refactoring on secondary ctor
Make custom handler that runs when caret is on secondary ctor
but do nothing because constructors don't have name.

It's neccessary because otherwise default handler will be ran
with proposition to rename class.

Also note, that we do not run this handler when
caret is on constructor's body or argument's list to allow renaming
inner declarations.

GitOrigin-RevId: 8e895f344161676dcb77ec043e262632bc512462