Enable resolving to the roots of the modules for Scratch files (PY-15818)
Scratch files are now located outside of the project, so foothold
PSI elements in them don't allow us to find their modules. For Scratch
files we explicitly add the ability to resolve to any of the modules
of the project, as we did earlier for the console runner.

(cherry picked from commit 6a2a3c2ee934e39cb7470528654572fc4bb2a2d1)