IDEA-140501 Send repo update events on a separate thread
Otherwise, if update() is called from
VcsRepositoryManager#getRepositoryForRoot (a Repository for the root
was not created yet => it is being created at once) other
messages from the message bus queue block the current thread from
returning the created GitRepositoryImpl instance, and moreover may
cause other GitRepositoryImpl() creations (because the repo map is
still not up-to-date).

A correct solution is to disable all git/hg actions while the repo map
is being updated, but it is a large change. On the other hand, sending
notification events on a separate thread shouldn't hurt.
(cherry picked from commit a88ad70)