Update binary skeletons path on every SDK update and clean up obsolete paths (PY-15499)
Sometimes paths to binary skeletons in jdk.table.xml used to become
corrupted or invalid. There used to be no way of fixing that since
the paths had been added there only once during the SDK set up. Now
we restore them every time the SDK gets updated. The paths should be
in the jdk.table.xml because they are the part of SDK paths and they
are required for indexing. Indexing contributors don't allow to
associate paths with modules.