Prefer submodules to '__all__' defined in the package (PY-14454)
We resolved references to attributes / submodules of a package to
'__all__' if we found the name of the attribute in '__all__'. That is
OK as a fallback if there is no appropriate submodule, but in other
cases we now resolve the reference to a submodule, not to '__all__'.

(cherry picked from commit 648fe87a05b590e5a8f975d9ffa9b041f428ab1c)