Replace list with set to resolve linear scan on list remove (IDEA-142174)
LocalFileSystemImpl has a collection of folders to be watched and closing
of project / closing of IDEA causes removal of these folders. As the collection is
an array list this operation is O(n) which causes problems when more than 20000 folders
are being watched and registered in that collection.
(cherry picked from commit 4ce804a) [r=Roman.Shevchenko]