MoveFileHandler#prepareMovedFile is called on a file before its directory is moved
Previously for the case of moved directory MoveFilesOrDirectoriesProcessor
invoked MoveFileHandler#prepareMovedFile on the files contained in
it only after that directory had been moved already. Such behavior
doesn't conform to the contract described in JavaDoc of the handler,
that claims that prepareMoveFile is called before any modification on
file is made. It caused significant burden on a developer who needed to
know original location of the moved file and had to retain this
information e.g. by using user data.