Resolve only explicitly imported submodules of imported packages (follow-up for PY-14454)
Now we try to filter resolve results for members of an imported module
by only explicitly imported results as we used to do for directories
only. The difference is we do it now for both directories and
__init__.py files and only if we're resolving a member in the same file
where the import for the module is located.

(cherry picked from commit f9d156ca8a529e5fdde472eed520166525a1bede)