explicit call of load is not required:
Application — was not used, only unrelated code (state store must not load components — it is component manager responsibility)

Project — useless, getMainStorage(false) will be called in any case in a lazy manner. Especially in case of directory based project.

Module — super impl was useless, because our impl in any case call getMainStorageData() Method is not removed — it is unclear who is reponsible to call "setOption(Module.ELEMENT_TYPE", so, code is not changed for now