Fixing type resolving PY-16476 and PY-3999.
Django Manager needs EP to obtain its parents (because it inhertis from Django-specific expressions), and this EP can't use type resolving (due to loop), so EP uses "resolve()" for reference while reference does not provide TypeEvalContext. So, we use codeInsight context to obtain list of parent types if context is not provided explicitly.