PY-11462 Fix inability to locate code sources when hitting breakpoints located at *.py library symlinks
On collecting Remote Library sources (sys.path locations) we may download one *.py file several times: there could be one or more symlinks of it and the original file (or we may not download the original file, because it could be not on the sys.path, but in /usr/share/pyshared for example). So having several versions of the same *.py file we need to store information about breakpoints in py debugger at one common key for what real path is well suited (which will be equal for original file and symlinks to it) (and this had been already done). But when breakpoint is hit we need to send absolute path of the file to point to the actually used version of it. If we send a real path we could point to a file that is not in sys.path and had not been even downloaded (as with /usr/share/pyshared).

(cherry picked from commit 8cefab7f780ff2c2657fc9e73bcbad87a01e67b9)