ExtensionPoint to provide Gradle plugin descriptions.
ApplyGradleDslPluginAction shows list of Gradle plugins and their
descriptions so that users can select one of them to add
"apply plugin <plugin-name>" to the gradle files.

This change is adding an extension point to provide the plugin
names and descriptions so that custom Gradle plugins (ex: android)
can constribute the additional Gralde plugins to the
ApplyGradlePluginAction.

This is required to address http://b.android.com/76734

Change-Id: Ib1970c2dd1f04a285d10b20ff01ee4d02edd3dca