Fix file length check
IJ uses the timestamp and file length to determine if two files
differ. In almost all cases, the timestamp will be sufficient
but there are some important use cases were timestamps are
manipulated inside of archives to ensure the archives only
differ when their actual contents differ.

The operation of reading the currently stored file size could
cause a read of file size from disk. That makes the up-to-date
check broken since it will be comparing the new file size to the
file size just read from disk, which should be the same as the
new file size.

https://github.com/JetBrains/intellij-community/pull/302/