[git] IDEA-142731 use correct VirtualFiles after refresh
If a file was deleted, we refreshed its parent (which is correct)
and returned the parent to the caller (which is wrong).
Instead, refresh as previously, but return only those VirtualFiles
which correspond to existent files.

(cherry picked from commit da5e17071e4b63a11a575e3de428c6229279be14)