Changed IncorrectPath test
Modified original test to not wait for events from the file
watcher. With the canonicalization, invalid paths are no longer
being sent to the file watcher, so the test would have waited
until TIMEOUT every time it looked for an event. Since it was
only doing this to ensure no events were reported, it was
expected behavior, but the test was not intended to test for
empty events, it was intended to ensure the file watcher still
worked correctly when it received and invalid path. The test now
tests that the file watcher doesn't crash when given a single
invalid path.

Added a second test that sets a valid watch root and an invalid
(non-existing) watch root. It tests to make sure the file
watcher works as expected for the valid watch root even when
an invalid root was sent to the watcher. I believe this closely
matches the original intention of the test.