Put path mapping storage behind an interface
Use a map to look up path mapping information. This makes the
lookup on the order of the length of the path (in components)
rather than on the order of number of paths that need mapping.

Added support (and a test) for the situation when a mapped path
is then remapped by the native file watcher.