[grails] pass VM options from Run Mvc Target dialog (IDEA-145433)
There was a choice:
- change MvcCommandExecutor to accept vmOptions parameter.
- change MvcCommand and make vmOptions as a field of MvcCommand.

The second approach was chosen because the first makes code less readable and GrailsCommandExecutor still needs `jvmArgs` parameter too. So this parameter was made a field of MvcCommand and it was removed from MvcFramework (and its inheritors) and GrailsCommandExecutor (and its inheritors).

Setters for MvcCommand fields were removed to avoid changing outside of MvcCommand.parse() method, so `vmOptions` parameters was added to it.

(cherry picked from commit 3f06dc343b2792736fc12741174c75f4d70f2c1d)