Add support for sending a module name to pydevd
Previously, the args for a subprocess spawned under debug with a module specified via '-m' would get rewritten incorrectly, resulting in the '-m' flag being split from the module name. This resulted in running a command similar to '/usr/bin/python -m /path/to/pydevd.py --someargs --file [module]', which doesn't work. This removes the '-m' from that command and adds a '--command' flag passed to pydevd, which results in correctly running as a module instead of a file.