IDEA-143358 Unregister authenticators on project dispose if they don't get unregistered for some reason
HTTP and SSH Authenticators hold a link to the Project,
while the GitXmlRpcHandlerService is application wide.

Normally, authenticators are unregistered on Git processTerminated,
or if an exception happens.
But if some reason this doesn't happen, we want to make sure that the
Project doesn't leak.

(cherry picked from commit 86eb25d344633ca424d4206e030371df4455c290)