Add ability to explicitly set strict=false
Due to the fact that strict mode is set if there is an 'idea.is.internal'
property, it's impossible to use internal mode and define an icon for action
group in plugin.xml while developing plugin. ActionManagerImpl fails for
plugin's icons if the strict mode enabled, but it is able to handle missing
icons itself.