IDEA-148739 [follow-up] Enable gutter icons scaling by default
Set a limit to the cache of icons of different sizes. When a gutter icon is scaled up/down as one is zooming, it can take about 50 values. There's no need to keep all those sizes in the cache. Moreover, the cache is rather used for a "common case", when an icon is scaled to meet different modes, like "standard", "presentation" w/ combination of possible "scaled" state. So, "5" should be enough for the cache limit.