IDEA-55508: compare revision numbers as strings
The history tab for the same file & revision pair should be reused,
but revision numbers can be provided from different places and thus
be of different implementations.
We can't force all VcsRevisionNumber implementations to provide equals
aware of other implementations, so just compare strings,
which is enough in this case.

Additionally, at any case, provide the standard equals & hashcode
for the TextRevisionNumber.