ES6. client debug:
source url is correct on meteor side.
our smart solution works — we don't need to setup mappings, we find script by content.

so... why we stopped at first line in the client/app.js?

Well... because there is a breakpoint in the server/app.js. As we set breakpoints only by file name, not but by file path, one VM breakpoint serves _2_ IDE breakpoints.
And there is a clear comment in the LineBreakpointManager — "we must not remove vm breakpoint - it is used for another ide breakpoints".
So, we just not remove VM breakpoint and it hit. And we stop.

Solution — fix "hasBreakProblem" and filter it out on suspended.

Well, we can implement specific solution for meteor, but... I am not going to fix debug for each tool on the planet. Generic solution is a bit complex, is a bit slower, but it is generic. Save our time and time of our users.