Fix multiprocessing debug with breakpoints in both main and child processes (PY-17092)
When the new process is created, the main thread in the new process already has the attribute 'pydevd_id', so the new thread doesn't get new id with its process number and the debugger loses access to both threads. Therefore we should update thread_id for every main thread in the new process.