Eliminate optimization in GroovyFileImpl.java to avoid StackOverflow.
We are seeing StackOverflowExceptions that are hard to reproduce.
There are no known steps to reproduce the problem consistently.
However we did encounter a sequence of stack frames which caused
a StackOverflowException with GroovyFileImpl.java

This CL eliminates what looks like an optimization that may have
caused the StackOverflowException we encountered.
Bug: http://b.android.com/199355

Change-Id: I5001f674470c2591633e3e008987d246cd0c05e6