Renamed NameDefiner to PyNamedElementContainer and changed semantics to match actual usage
NameDefiner.getElementNamed() wasn't used anywhere as an interface
method. The imprecise semantics of NameDefiner.iterateNames() resulted
in code that actually processes only, say, PyFile, but was written to
accept PyExceptPart or other NameDefiner descendants. That blocked us
from separating resolving named references and simple PSI children
crawls. It also resulted in single-resolve implementations of
getElementNamed() for import elements and Python files.

The only thing that didn't fit in the suggested scheme was
PyAssignmentToLoopOrWithParameterInspection that depends on the fact
that iterateNames() returns any PyElements for named elements, not only
PsiNamedElements, e.g. 'q[0]' for 'for q[0] in xs: ...'.