IDEA-150929 diff: set preferred focused component for a diff window
If getPreferredFocusedComponent in null, the DialogWrapper will not get focus. If this is the only window in IDE, Application.isActive() will return `false` and FocusManager will ignore all requests.

Before this fix, `myProcessor.requestFocus()` did nothing, as FrameWrapper would override this focus request with its own (on a component that was choosen by `IdeFocusManager.getFocusTargetFor()`, and not by `DiffRequestProcessor.getPreferredFocusedComponent()`).

This should also fix shortly visible "Loading..." panel in modal dialogs, as `myProcessor.updateRequest()` is now called immediately too.