Use multi-resolve for multiple values conditionally defined in an imported module (PY-18402)
Added PyFile.multiResolveName() that is a multi-resolve version of
getElementNamed(). The same method is added to PyImportedNameDefiner,
but there is no need in extracting a super inferface at the moment.
Finally, there is the new PyImportElement.multiResolve() in addition to
resolve().

Changed some single-resolve usages to multi-resolve acorss the codebase.
There are still many less important single-resolve usages that should
be converted to multi-resolve eventually. Single-resolve variants are
deprecated now.