a11y: Fix code completion interaction with nvda screen reader
This change improves the interaction between the code completion
window and the nvda screen reader.

* When code completion shows up, nvda reads "Code Completion" then the
first item selected in the list.

* When the selection changes in the completion list, nvda speaks the
selected item at it changes.

* When code completion is closed, nvda reads the inserted item text
and knows that the editor window has the focus (and follows caret
events).

Note that this change does not seem to improve the interaction with
the Jaws screen reader. Jaws never sees the code completion list,
probably because the list never actually gets the focus. Fixing that
issue would require a much more involved fix.

BUG=https://code.google.com/p/android/issues/detail?id=198011

Change-Id: I7976b115b0495d307c61b4b59e03952a33f08324