a11y: Set focus to error tooltip when invoked with "Ctrl-F1"
This makes browsing errors/tips of a single file fully accessible
via the keyboard:

* Use "F2" to go next error/tip location
* "Ctrl-F1" to access the error/tip text
* "ESC" to go back to the editor, and so on.

Most of the code is about threading a "requestFocus" boolean option
from the invocation location (ShowErrorDescriptionAction) to the
code that displays the tooltip (BalloonImpl). The reason we need to
thread this parameter is that ShowErrorDescriptionAction is the only
place where we know the current action was invoked from a key press
(as opposed to a mouse click, mouse over, etc.).

BUG=https://code.google.com/p/android/issues/detail?id=198205

Change-Id: I48657fe50dece8a541f12d572990eeea071a5550