We should really ignore Java extdirs
This was fixed in commit 4a20f3af by allowing ExtClassLoader to only
access the extension libraries of the JDK.

Later, upstream decided to move the extdir jars to the end of the
classpath (commit c7425ea2), and we integrated their patch with 15.0.3.

But we still have the problem of providers (such as BouncyCastleProvider)
scanning the classpath and running into incompatible classes towards the
end.

Bug: 201210
Change-Id: Ia85c2afa5b961077f178426e359f419eb95b2dd8