Adding product-build parameter to the plugin repository url.
The build entry in the AndroidStudioApplicationInfo.xml is defined as

<build number="__BUILD_NUMBER__" date="__BUILD_DATE__" apiVersion="AI-143.1821.5"/>

The apiVersion is manually updated every time we pull new
IntelliJ sources, i.e it indicates the version of IntelliJ sources we
used to the Android Studio. During the build process the __BUILD_NUMBER__
is replaced with a version number AI-143.12341234, where 12341234 is
the build number. When the apiVersion is available that will be used
to check the plugin compatibility and passed in as 'build' parameter
to plugin repository url. i.e to to make sure any plugins which are
developed for the IntelliJ version from which the Android Studio is
built, will also work. When the apiVersion is not defined in the xml
file, the build number will be used for the plugin compatibility and
the same is sent as 'build' parameter.

The newly added parameter will always pass the build number as the
'product-build' in the url and the url will now looks like ...

http://my-plugin-xml/?build=AI-143-1821&product-build=AI-143-12341234

This change also passes the correct apiVersion and buildNumber
of the platform to be updated. This happens when there is a platform
update available and we are querying the plugin repository to find
any plugin updates available targetting the platform to be updated.

Change-Id: I5169efaa8dc1bc01af3299f150862af89ed05980