Stamping aswb plugin and targeted product build version.
Patches plugin.xml to replace "SNAPSHOT" in the version with the
build number and preduct-build.txt to replace "PRODUCT_BUILD"
with the full build number.

Removing the idea-version entry from the plugin.xml as the filtering
based on product version is not working because the plguin compatibility
check always happens against the base platform version. Now there is no
compatibility filter in the plugin.xml and all that logic is moved to
the plugin repository server.

Change-Id: I623a4081b69ef7c8643c26082ef8e149727ef4f1