Update kryo lib with fix for Android detection
The kryo lib relied on checking if android.os.Process was in the
classpath. This would fail if layoutlib is in the classpath since it
would incorrectly report that is running in Android.
This has been merged into the kryo library in
https://github.com/EsotericSoftware/kryo/pull/404/ but this applies that
patch to 2.22

The jar file was built from the kyro-2.22 release tag by applying the
patch in this CL on top.

Change-Id: Iab57fe7f594d9a2562393a3aad7a6cef9c92694e