avoid deadlock condition in JavaDocumentationProvider (made possible after fix for IDEA-152628)
JavaDocumentationProvider now needs to read data from class documentation URL while generating URL for class'es method. This is performed under read action and can lead to a deadlock if a write action is pending, because built-server also needs a read action to generate response (new read action is not allowed to start if write action is pending). To avoid this situation, we use original javadoc URL (not proxied by built-in server) to fetch data internally