a11y: Make "Show Expression Type" screen reader friendly
Ensure the tooltip that is shown after expanding the expression
type has the focus so that the contents is read by screen readers.

Note the change in HintManagerImpl is more general, in the sense that
any tooltip shown via the HintManager will get the focus when
a screen reader is active.

BUG=https://code.google.com/p/android/issues/detail?id=198499

Change-Id: Ib55f8f07d4f20931f2c3baae9ae5346067a624f4