vcs: IDEA-153272 Don't discriminate unversioned and ignored files
When a file is moved and overwrites existing file (with the same name),
at first it is deleted in the new place.
This deletion would schedule a `git rm` operation, however there is
a logic which prevents from these files to be deleted in the case
of overwrite.
However this logic was not enabled for unversioned and ignored files.

filterOutUnknownFiles() is not needed anymore since unknown files
are not filtered in all cases.

(cherry picked from commit 783963b227a2450770c67c05ea344d51816038cf)