a11y: Make TreeTable component accessible
The TreeTable component has an accessibility issue related to the
column of the table that contain the tree: each cell in that column
is seen as a "tree" component (i.e. the full tree), whereas it makes
much more sense from an accessibility point of view to see each
cell as the tree *node* corresponding to that cell.

The fix is that instead of returning a TreeTableTree instance for
each rendered cell, we return a new component that wraps the
TreeTableTree for rendering purposes and the tree node component
for accessibility purposes. See the comment on
TableCellRendererComponent for a more detailed explanation.

This change should not affect the rendering of TreeTable
components (e.g. the "SDK Platforms" tab of the "Android SDK" settings
page). It should only affect what screen readers see for cells
in the column displaying the TreeTableTree.

Note: This is a somewhat subtle/tricky/non-trivial change, but it
is the best option I found to solve this problem after looking
into this for the last few days.

Change-Id: I463fefe4b596498c7644e87a76078cfa40684d24