a11y: Improve accessibility of ThreeStateCheckBox
There is currently no formal support for tri-state checkboxes in the
Java Accessibility API, this change attempts to workaround this
limitation.

The change improves a11y for ThreeStateCheckBox by

* Appending a state string ("checked", "partially checked",
  "not checked") to the accessible name of the checkbox so that
  screen readers say the state as part of the name of the component.
  This is not ideal, but at the end of the day, screen readers don't
  have a pre-defined string for "partially checked" for JAB
  applications -- they do for other accessibility APIs, such IA2.

* Changing the role from "checkbox" to "togglebutton" so that screen
  readers don't try say the implicit "not checked" state of
  checkbox.

Note:
Ideally, this would have been solved by using the
"AccessibleState.INDETERMINATE" state when a checkbox is partially
checked, but screen readers don't seem to pay attention to the
"indeterminate" state, so we had to resort to the workarounds
implemented in this change.

Change-Id: If6aee13fe27f0dc7489e0d033e45aed015a65bfc