PY-18892 Always use UTF-8 to interact with rest_formatter.py helper script
It's needed to avoid using fallback ASCII encoding when HTML produced
by docutils is written to sys.stdout, as we interact with the process
via pipe. It's better to be explicit about encoding in use, escpecially
because docstring may contain special characters such as math symbols.
See http://stackoverflow.com/questions/492483/setting-the-correct-encoding-when-piping-stdout-in-python
(cherry picked from commit dffea17)