a11y: Improve JBTable accessibility.
Address a subtle bug in JTable accessibility implementation by
customizing JBTable accessibility implementation.

The short story is that the AccessibleContext implementation
of JTable sometimes does the wrong thing when derived classes
depend on the ability to override the getCellRenderer method.
This leads to screen readers being unable to correctly report the
contents of certain cells.

The long story is documented in the comment of the
AccessibleJBTableCell class in this change.

Change-Id: I8b55cd4ca24ce5194540c0143780b9d699c46f0b