Fix NPE exception due to focus listener leak
The DarculaSpinnerUI installs a focus listener on the editor
component associated to a Spinner component. The problem is that
the listener was not removed in all cases, leading to leaks and
"NullPointerException" inside the focus listener handler.

The fix is to make sure the focus listener is removed when
"uninstallListeners" is called.

Change-Id: I8e42af32dd5e9c8658bbe273ca70720fae2ab910