a11y: Ensure action buttons show UI cues when they get the focus
Note: This change is not 100% related to accessibility.

In a previous change, we made ActionButtons focusable so that they
are easier to reach via keyboard (TAB key). To be "good citizens"
though, ActionButtons should show a "I got the focus" UI cue, as all
components do. This is the purpose of this CL.

Change-Id: Ic52c3f65738f3599ed7b61ed7f45fa0fbcb4549b