a11y: Activate editor when "return" is pressed in Project View.
Set the focus to the editor window when pressing the RETURN key on a
file node in the Project View tree -- as opposed to only activating
the editor frame while leaving the focus in the tree view.
This is similar to the way double-click works. This makes keyboard
activation more user friendly for a11y users.

BUG=https://code.google.com/p/android/issues/detail?id=199653

Change-Id: Icf9e9a756fffcc9af3c352592a7f27c9e1db5268