Tests: EditorFixture.waitForCodeAnalysisHighlightCount fail
We were using isErrorAnalyzingFinished(), but this is not enough.
Most tests type something on the editor, and then want to wait for the code
analyses to finish. But after typing, the Code Analyzer Daemon starts a timer,
and the analyses only starts after the user stops typing for a few time (about
300ms). Calling isErrorAnalyzingFinished() just after typing returns false.

isRunningOrPending() does what we need, but is package protected.

Change-Id: I73011105780d2daeaa2d236c5822fa7601521705