Implement authentication for the builtin webserver
Upon Android Studio start, it ensures there is a 'user.token'
file in the user's studio config dir (something like
 ~/.AndroidStudio20/config/user.token, only accessible by that user.
This file contains a cryptographic seed that is in text format.

The builtin webserver is modified to require basic authentication
in the form of user: '_token_', password: the contents of the 'user.token'
file above. Alternatively the token is allowed to be passed as a prefix
of the server path, see below why this is needed.

The BuiltInWebBrowserUrlProvider has been modified to launch
the urls with a prefix of the token in the path of the url.
This is needed as while most browsers support urls in the form of
'http://user:pass@host', Internet Explorer does not.

The SSH and HTTP git plugins have been modified to pass the right
username and password combination to the XmlRpc apis for Basic Auth.

Users who want to access the builtin webserver from the command-line
can run something like:

   curl -D -  -u _token_:`cat user.token` \
    http://localhost:63342/api/file?file=path-to-file

Cherry-picked from b37ba0282b74a9612c07ee7a2f6660988345b3d1,
  d1d95327414bfb93ef158b8994d040c669f7664d,
  b9fcd6018d152d640540e6c1a90a897295f00fa5,
  43dbe1b5d14554d90d7faa07b44bfd3f528ebc79,
  7a5cbd2e1089d75fe02833513f5a7afed9a57d47,
  f0ac2ae46643edc902df729e75c2feae4f910405,
  996f0e5d9c54ae5c6c6f65a94eac95f152cc3f71

Change-Id: I9dfdccbd0616a26762705321364aa57e13db2ccb
Fixes: 137372765