Adds support for a prebuild step in Studio.
The android plugin now contains a build.xml that is used
to trigger a prebuild step when compiling studio.

This xml defines the prebuild task that is called from IJ
and from ant when building Android Studio. The prebuild
steps is used to build external dependencies needed by
studio which currently are built using gradle.

Change-Id: Ibe673c30c19c8a163f15bc11247fd2e91b39b592