a11y: Ensure navigation bar works nicely with screen readers
The navigation bar needs to allow navigation bar *items* to receive
the focus to be screen readers friendly. Allowing navigating between
navigation bar items with the LEFT/RIGHT cursor keys is not enough,
because screen readers rely on focus events to announce UI components
type, label, etc. So, when handling LEFT/RIGHT, etc. we need to set
the focus to the corresponding nav bar item component.

Although this sounds like a simple change, it turns out this was
somewhat tricky, as there were a few spots in the navigation bar
code that assumed that "navigation bar is active" is equivalent to
"navigation bar component is the focus owner". This is not the case
anymore, so we had to update the code to loosen this assumption.

Change-Id: Ic9d9c9a0a3f7496d07a7bb380ffde7d577e57a07