2nd attempt to fix the missing dispose of editor.
Whenever a project is closed that has an open inspector we get
an error about a missing dispose of an editor. This is an attempt
to fix this problem.
This time: make the editor invisible immediately and delay the
component removal.
We should try to use this for a while before accepting this into
2.2

Change-Id: Ibfba899f00254a0addd4a1c969363e13de772684