a11y: Make error messages easier to navigate/read.
A previous change partially addressed the issue of making error
messages readable with screen readers, but the user interaction with
keyboard navigation was still sub-par. With this change, we ensure
error message are always "non editable" if a screen reader is active.

This ensure that nodes in the tree view of the "Messages" tool window
appear as simple labels. The unfortunate side effect is that any
hyper-link inside error message won't be accessible to visually
impaired users. This is not such a big deal at this point, as most
(all?) screen readers don't seem to support hyperlinks with the
Java Access Bridge.

BUG=https://code.google.com/p/android/issues/detail?id=208638

Change-Id: I3ce1e1c28cdf0c7cda610587dc76aa428ce1ac70