a11y: Ensure Shortcut edit field are focus traversable
By default, when the "ShortcutTextField" receives the focus, it
listens to *all* keystroke events and captures them as the assigned
keystroke. The problem is that TAB/Shift-TAB are also captured, hence
preventing the focus from leaving the field.

Given Tab/Shift-TAB are valid keystrokes to be captured, we solve
this problem by 1) allowing the ShortcutTextField to capture the
TAB/Shift-TAB keystrokes and 2) *then* allowing those keystrokes to
move the input focus to the next/previous field.

This is somewhat sub-optimal user experience, but addressing this
issue more cleanly would involve churning/revamping the user
interaction model.

Also, this change makes the validation messages focusable, so that
TAB/Shift-TAB navigates through them.

BUG=https://code.google.com/p/android/issues/detail?id=198489

Change-Id: I75d48e235d5fa9b86eae93735cf5c99cc5253f00