PY-17815 EA-67335 Don't save PSI reference in quick fix, save its class and find it later
Also I prohibited showing the hint when underlying PSI element was
changed but revived by SmartPointer when these old and new elements
have different names.

I'm not yet sure whether it's possible to get rid of PSI reference
in the quickfix completely since there might be multiple references
on the same element and we don't want to suppress the hint because
some other unrelated references are resolved successfully (PY-3167).