a11y: Fix processing of LEFT/RIGHT keys in TabLabel
When a TabLabel has the focus, the LEFT/RIGHT keys should allow
selecting the previous/next tab. There was an attempt to implement
that behavior made earlier, but it didn't work in all cases, because
TabLabels can be re-ordered and made invisible. This change fixes
the code to use an API specifically built for this purpose.

Change-Id: Ibdda6c5012b604d32e88cfbd9b2a8d7c180bac4d