Integrate analytics-library into Android Studio
ensures that analytics tracker & publisher are initialized,
relies on user to have opt-ed-in in Android Studio UI.

We want the Android Studio usage tracking to be available
from the platform, so this CL adds the entire metrics code
into the platform itself. The android plugin will depend on
the same libraries. We expect these libraries to be bundled
inside the android plugin in the IntelliJ distribution.

Change-Id: Id46e8175c5a827e9e87ca72a59117e9ac007a267