[vcs-log] get rid of VcsRootsRegistry
Since hash map stores data in a file for a specific set of roots, we do not need root indexes to be consistent in a project, only in a specific set of roots. This means we do not need to store roots on disk at all. This solves various issues with having another PE.
(cherry picked from commit 87145fa)