PY-19975 Check all resolved symbols in __init__.py to find shortcut name for element
The root of the problem however is that PythonPathCache often contains
relative qualified names as keys. It leads to the situation where
QFrame module is resolved to the corresponding files both in PyQt and
PySide. We should probably either omit such resolve result in
PythonPathCache entirely or use absolute qualified names for them as
keys.