Debugger stopping in all threads after resuming from a breakpoint (PY-20158)
Sometimes it isn't possible to suspend thread, which is sleeping or waiting on lock. This thread is trying to suspend, but it's not suspended yet. When user is sending "Resume" command, we don't resume this thread, because it isn't suspended. After that, when this thread stops waiting, it doesn't know anything about Resume command and it's trying to suspend.
So, we need to send Run command to all threads (suspended and non-suspended) and we should do it without using internal command.
(cherry picked from commit 79b865d)