IDEA-65721 Git Sign-off
Add a checkbox to the commit dialog which calls `git commit -s`
thus adding "Signed-off-by" suffix to the commit message.

There was an idea to add a toggle button which wouldn't call "-s",
but modify the commit message in place. However it appeared that this
process could cause confusion in several cases, e.g.:
* user usually signs commits in this project;
* he opens the commit dialog, the signature is already there at the end
  of the previous commit message;
* the commit message is selected to ease writing new commit message;
* user starts typing, everything gets erased =>
* should the sign-off suffix be added or not? The button is still
  toggled, because it is a project wide control.
The checkbox is a bit less neat option, but lets avoid such confusions.

The state of the checkbox is remembered in Git project settings.

Several methods became non-static to avoid passing 3-4 parameters
which can be easily retrieved from the class fields.