Workaround dx/jack bug that include null local vars
Both Jack and dx emit a local variable table that includes method
parameters, but with null parameter names. Until API 23, these variables
weren't reported by the VM when requesting the local variables in a
frame, but API 24 does report them. It doesn't make much sense to report
a value for a method parameter with no name associated to it, so this CL
filters those out.

b/30673627
https://code.google.com/p/android/issues/detail?id=219146

Change-Id: I35fa85953d02b4f482f84524d30f4d8b4470da44