PY-20058 Fix the check that we should suppress E402 in Jupyter notebooks
So that it ignores preceding comments and imports and, overall,
mutes this error not only on the first import statement in a cell.

Also I accidentally used the interface PyImportStatement instead of
PyImportStatementBase before, so, the previous fix didn't handle the
original case with "from __future__ import print_function".
(cherry picked from commit 10fffd2)