IDEA-159965 better solution to display email link in log details
Name being a link to email has two drawbacks:
* not clear what is it link to;
* no way to copy email.

The solution is to write the e-mail in the standard form after the name,
and also make it a link to mailto.

If committer is different from author, the line gets too wide,
so let's move it to a separate line, aligning with the first one.