Report Studio crashes to crash backend instead of GA
A previous CL changed the exception tracking backend from GA to crash.
This CL makes all crash reports go to (Google) crash as well. In the
earlier scheme, we were sending a single event per crash to GA, but now,
all the crashes are combined into a single report that has two custom
fields: numCrashes and crashDesc that reflect the total number of
crashes in this report, and the description of all of the crashes
combined into a single string respectively.

Change-Id: Ieee2a127c3e5c7429c81b09f9ac3228925bf122b