Leak checker: ignore the ProxyUtil cache
InternalAndroidModelViewTest creates some proxies of MyInterfaceImpl
which leak the enclosing class and therefore have access to
AndroidGradleTestCase.myModules and the soon-to-be-disposed project.
They get added to the soft-value ProxyUtil cache, and stay there until
the GC collects the values. The leak checker will flag these as leaks,
hence the flaky behavior.

Change-Id: Id85ba2f859d357b175822f76e1384516bf0834b1