git: fetch multiple roots in parallel
Submit several tasks using the standard bounded executor, then wait for their completion.

Execute tasks under a common progress indicator, to make sure that all tasks are cancelled when one of them is.

The number of threads to execute is to be defined more precisely. Initially the number of cores is taken, but it doesn't seem to be obviously good enough default for a networking operation.

Make two implementations of RepoResult: one for fair per-remote result, another - for general error happening during the fetch. This could be solved by operating the pair of GitRepository & GitRemote, but the latter appeared to be less handy.

This fixes IDEA-160208