Send PerformanceWatcher reports to Google crash
The IDE already includes a PerformanceWatcher that detects when the IDE
is stalled and dumps all threads at that stage. This CL listens to that
event add does two things:
 - reports how long the IDE was stalled (via metrics pipeline)
 - reports the thread dump (via crash pipeline)

We don't want to do additional work when the IDE is already stuttering,
so both of these do minimal work at the time of the event: the time is
saved in the metrics offline db, while the path to the thread dump is
saved to the new thread dumps database (just a list of paths to thread
dumps). When the IDE is started the next time, it reports a max of 10
thread dumps from the previous iteration.

Change-Id: I8781533bab4681401ec689607bf383907dcb5e9b